requirements.txt
crewai-tools[mcp]
crewai
mcp
python-dotenv
pandas
textblob